# Remove usage of `rand`.